PRDBench/13/evaluation/expected_requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
scikit-surprise>=1.1.3
jieba>=0.42.1
gensim>=4.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.0
mysql-connector-python>=8.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
mlflow>=2.7.0
pytest>=7.4.0
python-multipart>=0.0.6
pydantic>=2.4.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
scipy>=1.10.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pillow>=10.0.0